Removing duplicate tests
